requirements.txt
datasets==2.14.5
diffusers==0.25.1
numpy==1.23.5
pandas==2.0.2
transformers==4.36.2
wandb==0.15.11
accelerate==0.20.3
blobfile==2.1.1
scipy==1.10.1
seaborn==0.12.2
tqdm==4.66.1
SoundFile==0.10.2
librosa==0.9.2
ftfy==6.1.1
progressbar==2.5
torchlibrosa==0.1.0
einops==0.7.0
PyYAML==6.0.1
gradio==3.42.0
scikit-image==0.21.0
absl-py==2.0.0
ssr-eval==0.0.7
wget==3.2
llvmlite==0.41.0
scikit-learn==1.3.1
h5py==3.9.0
regex==2023.8.8
webdataset==0.2.48
braceexpand==0.1.7
mpi4py==3.1.5
laion-clap==1.1.4
matplotlib
pydub==0.25.1
pyloudnorm==0.1.1
python-dotenv==1.0.0